Server controller, router, model created
Successfully call data
