+++ server/router.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+const controller = require('../db/controller.js');
+
+//route all get requests from song endpoint
+router.get('/song', controller.song.get);
+
+module.exports = router;
